[FIX/UPDATE] Finish OO approach, new test (demo_functions), fixed issues
